[fe] refactor (#95) gql query string을 별도 폴더로 분리
