+++ client/src/components/routers/DriverRouter.tsx
@@ -3,7 +3,7 @@ import { Redirect, Route } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 import { ActivityIndicator } from 'antd-mobile';
 import styled from 'styled-components';
-import { IS_DRIVER_AUTHORIZED } from '@utils/authGQLString';
+import { IS_DRIVER_AUTHORIZED } from '@queries/authQueries';
 
 const DriverRouter: React.FC<any> = ({ component, ...rest }) => {
   const { loading, error, data } = useQuery(IS_DRIVER_AUTHORIZED);
