+++ client/src/queries/authQueries.ts
