+++ client/src/queries/driver/driverSignin.ts
@@ -1,0 +1,10 @@
+import { gql } from '@apollo/client';
+
+export const SIGNIN_DRIVER = gql`
+  mutation DriverSignin($id: String!, $password: String!) {
+    driverSignin(id: $id, password: $password) {
+      success
+      message
+    }
+  }
+`;
